Optimisation of Nystrom computation
update: MatrixPower in Util.py. The pinv initial step is reoved as it is
        useless and computationnaly expensive.
add: MatrixPowerh in Util.py. If the matrix is hermitian, eigh claims
     returning a unitary matrix, which can be inversed through a transpose
     step (less expensive than standard inversion).
update: in Nystrom.py. Calls to matrixPower is replaced by calls to
        MatrixPowerh.
update: eigpsd in Nystrom.py. Creation of submatrices A and b speeded up.
update: eigpsd in Nystrom.py. Creation and use of lmbdaN secured and speeded up.